Rename SolverMobile to Solver
