+++ packages/scrabble-solver/src/components/Solver/Solver.module.scss
@@ -1,6 +1,6 @@
 @import 'styles/mixins';
 
-.solverMobile {
+.solver {
   --spacing: var(--spacing--xl);
 
   position: relative;
