+++ packages/scrabble-solver/src/components/Solver/components/ApplyButton/ApplyButton.module.scss
