+++ packages/scrabble-solver/src/components/Solver/components/ApplyButton/index.ts
