+++ packages/scrabble-solver/src/components/Solver/components/SolveButton/index.ts
