+++ packages/scrabble-solver/src/components/Solver/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Solver';
