--- packages/scrabble-solver/src/components/SolverMobile/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SolverMobile';
